chore: Move AsyncTest out from public package
