+++ Package.swift
@@ -36,8 +36,7 @@ let amplifyTargets: [Target] = [
         dependencies: [
             "Amplify",
             "CwlPreconditionTesting",
-            "AWSPluginsCore",
-            "AmplifyAsyncTesting"
+            "AWSPluginsCore"
         ],
         path: "AmplifyTestCommon",
         exclude: [
@@ -274,22 +273,8 @@ let analyticsTargets: [Target] = [
     )
 ]
 
-let asyncTestingTargets: [Target] = [
-    .target(
-        name: "AmplifyAsyncTesting",
-        dependencies: [],
-        path: "AmplifyAsyncTesting/Sources/AsyncTesting",
-        linkerSettings: [.linkedFramework("XCTest")]
-    ),
-    .testTarget(
-        name: "AmplifyAsyncTestingTests",
-        dependencies: ["AmplifyAsyncTesting"],
-        path: "AmplifyAsyncTesting/Tests/AsyncTestingTests"
-    )
-]
-
 let targets: [Target] = amplifyTargets + apiTargets + authTargets + dataStoreTargets + storageTargets +
-                        geoTargets + analyticsTargets + asyncTestingTargets
+                        geoTargets + analyticsTargets
 
 let package = Package(
     name: "Amplify",
@@ -322,10 +307,6 @@ let package = Package(
         .library(
             name: "AWSPinpointAnalyticsPlugin",
             targets: ["AWSPinpointAnalyticsPlugin"]
-        ),
-        .library(
-            name: "AmplifyAsyncTesting",
-            targets: ["AmplifyAsyncTesting"]
         )
     ],
     dependencies: dependencies,
